Add support for some more Attributes
AnnotationDefault
EnclosingMethod
LineNumberTable
LocalVariableTable
"Markers" (Deprecated, Synthetic)

+++ src/main/java/org/stummi/jadis/reader/attribute/AnnotationElementValueReader.java
@@ -38,7 +38,7 @@ public class AnnotationElementValueReader implements
 			throws IOException {
 		char c = (char) jadis.readUnsignedByte();
 		if (!elementTypeMap.containsKey(c)) {
-			throw new IOException("unknown annotation value type: " + c);
+			throw new IOException("unknown annotation value type: " + (int) c);
 		}
 
 		return jadis.readElement(elementTypeMap.get(c));
